# maize/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "API documentation for Product endpoints",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/recommendations": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get list of all recommendations",
                "description": "Display a listing of the recommendations.",
                "operationId": "b50096a7b015b6eb5e37f889d4d86223",
                "responses": {
                    "200": {
                        "description": "List of recommendations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Create a new recommendation",
                "description": "Store a newly created recommendation in storage.",
                "operationId": "65096d023ff0777c624e8f51ee0d4ceb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "field_id",
                                    "recommendation_date",
                                    "recommendation_type",
                                    "message"
                                ],
                                "properties": {
                                    "field_id": {
                                        "type": "string"
                                    },
                                    "recommendation_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "recommendation_type": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recommendation created"
                    }
                }
            }
        },
        "/recommendations/{recommendation}": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get a specific recommendation",
                "description": "Display the specified recommendation.",
                "operationId": "15ed7af15fa9a3048e52b716fcbb2482",
                "parameters": [
                    {
                        "name": "recommendation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation details"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Update a recommendation",
                "description": "Update the specified recommendation in storage.",
                "operationId": "70711eeefa85f5cc1f71543d92c65da9",
                "parameters": [
                    {
                        "name": "recommendation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "recommendation_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "recommendation_type": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recommendation updated"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Delete a recommendation",
                "description": "Remove the specified recommendation from storage.",
                "operationId": "bf76b8843f1ef30c866ca4e5c6253dba",
                "parameters": [
                    {
                        "name": "recommendation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recommendation deleted"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    }
                }
            }
        },
        "/sensors": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Get list of all sensors",
                "description": "Display a listing of the sensors.",
                "operationId": "bd490465cc6a8216bb89455e149f4dd6",
                "responses": {
                    "200": {
                        "description": "List of sensors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Create a new sensor",
                "description": "Store a newly created sensor in storage.",
                "operationId": "ad22243e7d669a455d43db153d3b9e90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "field_id",
                                    "sensor_type",
                                    "installation_date",
                                    "status"
                                ],
                                "properties": {
                                    "field_id": {
                                        "type": "string"
                                    },
                                    "sensor_type": {
                                        "type": "string"
                                    },
                                    "installation_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sensor created"
                    }
                }
            }
        },
        "/sensors/{sensor}": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Get a specific sensor",
                "description": "Display the specified sensor.",
                "operationId": "60dc5af5fb5d87ae03c74f6ffed53968",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor details"
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Update a sensor",
                "description": "Update the specified sensor in storage.",
                "operationId": "1d70ea6b51af8d33e8124a8e8eb06b4d",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sensor_type": {
                                        "type": "string"
                                    },
                                    "installation_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sensor updated"
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Delete a sensor",
                "description": "Remove the specified sensor from storage.",
                "operationId": "2cac48377a74621fd696b50a2dcb3218",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sensor deleted"
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}